Week_6/demo-2-code-review-orchestra/requirements.txt
click>=8.1.0
rich>=13.0.0
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
python-dotenv>=1.0.0
httpx>=0.27.0
